requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
openpyxl>=3.0.0
Pillow>=9.0.0
pytesseract>=0.3.10